Added better docs for cellRangeRenderer
