+++ source/Grid/Grid.js
@@ -10,6 +10,7 @@ import getUpdatedOffsetForIndex from '../utils/getUpdatedOffsetForIndex'
 import raf from 'raf'
 import shallowCompare from 'react-addons-shallow-compare'
 import updateScrollIndexHelper from './utils/updateScrollIndexHelper'
+import defaultCellRangeRenderer from './defaultCellRangeRenderer'
 
 /**
  * Specifies the number of miliseconds during which to disable pointer events while a scroll is in progress.
@@ -721,74 +722,4 @@ export default class Grid extends Component {
 
     this._invokeOnScrollMemoizer({ scrollLeft, scrollTop, totalColumnsWidth, totalRowsHeight })
   }
-}
-
-function defaultCellRangeRenderer ({
-  cellCache,
-  cellRenderer,
-  columnSizeAndPositionManager,
-  columnStartIndex,
-  columnStopIndex,
-  isScrolling,
-  rowSizeAndPositionManager,
-  rowStartIndex,
-  rowStopIndex
-}) {
-  const renderedCells = []
-
-  for (let rowIndex = rowStartIndex; rowIndex <= rowStopIndex; rowIndex++) {
-    let rowDatum = rowSizeAndPositionManager.getSizeAndPositionOfCell(rowIndex)
-
-    for (let columnIndex = columnStartIndex; columnIndex <= columnStopIndex; columnIndex++) {
-      let columnDatum = columnSizeAndPositionManager.getSizeAndPositionOfCell(columnIndex)
-      let key = `${rowIndex}-${columnIndex}`
-      let renderedCell
-
-      // Avoid re-creating cells while scrolling.
-      // This can lead to the same cell being created many times and can cause performance issues for "heavy" cells.
-      // If a scroll is in progress- cache and reuse cells.
-      // This cache will be thrown away once scrolling complets.
-      if (isScrolling) {
-        if (!cellCache[key]) {
-          cellCache[key] = cellRenderer({
-            columnIndex,
-            isScrolling,
-            rowIndex
-          })
-        }
-        renderedCell = cellCache[key]
-      // If the user is no longer scrolling, don't cache cells.
-      // This makes dynamic cell content difficult for users and would also lead to a heavier memory footprint.
-      } else {
-        renderedCell = cellRenderer({
-          columnIndex,
-          isScrolling,
-          rowIndex
-        })
-      }
-
-      if (renderedCell == null || renderedCell === false) {
-        continue
-      }
-
-      let child = (
-        <div
-          key={key}
-          className='Grid__cell'
-          style={{
-            height: rowDatum.size,
-            left: columnDatum.offset,
-            top: rowDatum.offset,
-            width: columnDatum.size
-          }}
-        >
-          {renderedCell}
-        </div>
-      )
-
-      renderedCells.push(child)
-    }
-  }
-
-  return renderedCells
 }
